v4l-gst: Remove no-reorder=true from omxh264dec in the pipeline setting
It seems that the behavior of this option on RZ/G3E VLP 1.0.0 + Chromium
v132 is inverted from RZ/G2L VLP 3.0.6 + Chromium v109.
On RZ/G2L, if B-frames are present in a video track, display order of
video frames was not expected, this is the reason we added
`no-reorder=true`.
On RZ/G3E, this hack seems no longer needed.

It might be caused by Chromium since the following command gets same
result on both environment:

```
$ gst-launch-1.0 filesrc location=sample.mp4 \
  ! qtdemux ! h264parse ! omxh264dec no-reorder=true ! waylandsink
```

Signed-off-by: Takuro Ashie <ashie@clear-code.com>